requirements.txt
pythonnet>=3.0.1
requests>=2.31.0
pyodbc>=4.0.39
python-dotenv>=1.0.0
tenacity==9.1.2
tqdm>=4.65.0
psycopg2-binary>=2.9.9
